# src/main/resources/static/openapi.yaml
# 놓치지마 서버 API — OpenAPI 스펙
# 원천 계약: 루트 API.md (v0.2, 2026-08-28) — API.md가 바뀌면 이 파일도 함께 갱신한다.
# 열람: /swagger-ui.html (로컬 http://localhost:5000/swagger-ui.html)
openapi: 3.0.3
info:
  title: 놓치지마 서버 API
  version: "0.2"
  description: |
    출근길 출발 타이밍 알림 미니앱(앱인토스)의 파트너 서버 API.
    - 원천 계약 문서: 저장소 루트 `API.md` (v0.2) — 이 스펙은 그 계약의 Swagger 표현이다.
    - 유저는 전체 정류장/역(서울·경기 버스 + 수도권 지하철)에서 검색·선택한다 (2026-08-28 오너 결정, 명세서 §9).
    - 구현 상태 (2026-09-03): §1·§2·§3·§5·§6·§7 구현·배포됨 (§6·§7은 NCP 키 설정 전까지 503 → 클라이언트 폴백). §4 푸시는 발송 파이프라인 구현·dry-run 동작 중(템플릿 검수 후 라이브). §8 토스 로그인 구현됨 — mTLS 인증서(TOSS_MTLS_KEYSTORE) 설정 전까지 /auth는 503, 요청 인증은 dev 고정 유저 폴백. 인증서 설정 후에는 Bearer 검증→userKey로 유저별 저장.
servers:
  - url: https://catchmyride.hansw.dev
    description: 운영 (Oracle VM + Caddy HTTPS)
  - url: http://localhost:5000
    description: 로컬 Docker
security:
  - tossBearer: []
tags:
  - name: commute-routes
    description: 통근 경로 (유저당 최대 5개 — 출근·퇴근 등 라벨 구분, 2026-09-05 §1-2 추가)
  - name: commute-setting
    description: (레거시) 유저당 1개 시절의 통근 설정 — 첫 경로에 대한 어댑터로 유지
  - name: arrivals
    description: 라이브 뷰 도착 정보 (FR-501/502)
  - name: feedback
    description: 탑승 피드백 — North Star 측정 (FR-601)
  - name: stops
    description: 정류장/역 전체 검색 (온보딩 FR-102·103)
  - name: map
    description: 지도 미리보기·주소 검색 프록시 (온보딩 집 위치, FR-101 보완 — 2026-09-02 §6·§7 추가)
  - name: auth
    description: 토스 로그인 토큰 교환 (S-3/C-1 — 2026-09-03 §8 추가, mTLS 인증서 설정 전까지 503)

paths:
  /api/v1/commute-routes:
    get:
      tags: [commute-routes]
      summary: 통근 경로 목록 (생성순 — 첫 경로가 기본)
      responses:
        "200":
          description: 경로 목록
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommuteRoutesResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [commute-routes]
      summary: 통근 경로 생성 (최대 5개)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommuteRouteRequest" }
      responses:
        "201":
          description: 생성된 경로 (id는 서버 발급)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommuteRoute" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /api/v1/commute-routes/{routeId}:
    parameters:
      - name: routeId
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [commute-routes]
      summary: 통근 경로 수정
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommuteRouteRequest" }
      responses:
        "200":
          description: 수정된 경로
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommuteRoute" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/SettingNotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      tags: [commute-routes]
      summary: 통근 경로 삭제
      responses:
        "204": { description: 삭제 완료 }
        "404": { $ref: "#/components/responses/SettingNotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /api/v1/commute-setting:
    get:
      tags: [commute-setting]
      summary: (레거시) 첫 경로의 설정 조회
      responses:
        "200":
          description: 첫 경로의 통근 설정
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommuteSetting" }
        "404": { $ref: "#/components/responses/SettingNotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      tags: [commute-setting]
      summary: (레거시) 첫 경로 갱신 — 없으면 '출근' 라벨로 생성
      description: 다중 경로(§1-2) 이전 클라이언트 호환용 어댑터.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommuteSetting" }
      responses:
        "200":
          description: 저장된 객체 반환
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommuteSetting" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      tags: [commute-setting]
      summary: (레거시) 모든 경로 삭제 — 재온보딩 (알림도 중지)
      responses:
        "204": { description: 삭제 완료 }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /api/v1/arrivals:
    get:
      tags: [arrivals]
      summary: 라이브 뷰 도착 정보
      description: |
        해당 경로 설정의 정류장·노선 기준으로 서버가 공공 API(캐시 TTL 15~30초)를 조회해 내려준다.
        클라이언트는 15~30초 간격 폴링 (FR-204). 오프라인 시 마지막 응답 + fetchedAt 유지 표시 (NFR-07).
      parameters:
        - name: routeId
          in: query
          required: false
          schema: { type: string }
          description: 대상 경로 — 생략 시 첫 경로 (레거시 호환)
      responses:
        "200":
          description: 도착 정보 목록 (secondsToArrival 오름차순, null 마지막)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArrivalsResponse" }
        "404": { $ref: "#/components/responses/SettingNotFound" }
        "503":
          description: 공공 API 장애 — 클라이언트는 "실시간 정보 없음" 표시 (NFR-03)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /api/v1/boarding-feedback:
    post:
      tags: [feedback]
      summary: 탑승 피드백 원탭 제출
      description: 푸시 수신 후 "탔어요/놓쳤어요" — 같은 notifiedDate 재제출 시 마지막 값으로 갱신.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BoardingFeedbackRequest" }
      responses:
        "201":
          description: 기록 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  recorded: { type: boolean, example: true }
        "400":
          description: 그날 알림 발송 이력이 없음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /api/v1/stops/search:
    get:
      tags: [stops]
      summary: 정류장/역 전체 검색
      description: |
        온보딩 FR-102용. 서울 버스(TOPIS 정류소명 검색) + 경기 버스(GBIS) + 수도권 지하철(서버 내장 역 목록)
        세 소스를 병렬 조회해 통합한다. 제한된 목록 제시 금지 (명세서 §9, 2026-08-28 오너 결정).
        이 엔드포인트만 온보딩 중 통근 설정 없이도 호출 가능 (404 SETTING_NOT_FOUND 없음).
      parameters:
        - name: query
          in: query
          required: true
          schema: { type: string, minLength: 2 }
          example: 여의도
      responses:
        "200":
          description: 소스당 최대 10건, 정확 일치 우선
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StopSearchResponse" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /api/v1/stops/routes:
    get:
      tags: [stops]
      summary: 정류장 경유 노선 조회
      description: 온보딩 FR-103용 — 선택한 정류장에서 타는 노선 선택지.
      parameters:
        - name: type
          in: query
          required: true
          schema: { $ref: "#/components/schemas/StopType" }
        - name: stopId
          in: query
          required: true
          schema: { type: string }
          example: "19284"
      responses:
        "200":
          description: 노선 목록
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StopRoutesResponse" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /api/v1/map-preview:
    get:
      tags: [map]
      summary: 지도 미리보기 (NCP Static Map 프록시)
      description: |
        온보딩/재설정 집 위치 단계의 지도 표시용 — NCP 키는 서버에만 보관.
        마커는 클라이언트가 오버레이하므로 순수 지도만 반환. 인증 불필요.
        키 미설정·상류 장애는 503 — 클라이언트가 OSM 타일로 자동 폴백.
      security: []
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, minimum: -90, maximum: 90 }
          example: 37.5219
        - name: lng
          in: query
          required: true
          schema: { type: number, minimum: -180, maximum: 180 }
          example: 126.9245
        - name: w
          in: query
          description: 픽셀 폭 — 서버에서 1~1024로 클램프
          schema: { type: integer, default: 708 }
        - name: h
          in: query
          description: 픽셀 높이 — 서버에서 1~1024로 클램프
          schema: { type: integer, default: 320 }
        - name: level
          in: query
          description: 줌 레벨
          schema: { type: integer, default: 16 }
      responses:
        "200":
          description: 지도 이미지
          content:
            image/png:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "503":
          description: NCP 키 미설정 또는 상류 장애 — 클라이언트 OSM 폴백
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/v1/geocode:
    get:
      tags: [map]
      summary: 주소 검색 — 지오코딩 (NCP Geocoding 프록시)
      description: |
        집 위치를 GPS 대신 주소로 등록 (FR-101 보완, GPS 부정확 대응). 인증 불필요.
        최대 10건, 결과 없으면 빈 배열(에러 아님).
      security: []
      parameters:
        - name: query
          in: query
          required: true
          schema: { type: string, minLength: 2 }
          example: 여의공원로 101
      responses:
        "200":
          description: 주소 후보 목록
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    maxItems: 10
                    items:
                      type: object
                      required: [roadAddress, jibunAddress, latitude, longitude]
                      properties:
                        roadAddress: { type: string, example: 서울특별시 영등포구 여의공원로 101 }
                        jibunAddress: { type: string, example: 서울특별시 영등포구 여의도동 23 }
                        latitude: { type: number, example: 37.5219 }
                        longitude: { type: number, example: 126.9245 }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "503":
          description: NCP 키 미설정 또는 상류 장애
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/v1/auth/mode:
    get:
      tags: [auth]
      summary: 인증 모드 조회 (§8-0) — 클라이언트가 토스 로그인/익명 키 중 선택하는 기준
      security: []
      responses:
        "200":
          description: 토스 로그인 라이브 여부 — false(또는 조회 실패)면 클라이언트는 익명 키(§8-3, Bearer anon:{hash}) 사용
          content:
            application/json:
              schema:
                type: object
                required: [tossLoginAvailable]
                properties:
                  tossLoginAvailable: { type: boolean, example: false }

  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 토스 로그인 — appLogin 인가 코드를 토큰으로 교환 (§8-1)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorizationCode, referrer]
              properties:
                authorizationCode: { type: string, description: 클라이언트 appLogin() 결과 (유효 10분) }
                referrer: { type: string, enum: [DEFAULT, SANDBOX] }
      responses:
        "200":
          description: 발급된 토큰 — 이후 모든 요청의 Bearer로 사용
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503":
          description: 토스 로그인 미설정(mTLS 인증서 대기) 또는 토스 서버 장애 — 클라이언트는 헤더 없이 동작(dev 폴백)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: accessToken 갱신 (§8-2) — accessToken 1시간·refreshToken 14일
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: 새 토큰 쌍
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401":
          description: refreshToken 만료/무효 — 클라이언트는 appLogin 전체 로그인으로 폴백
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: 토스 로그인 미설정 또는 토스 서버 장애
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

components:
  securitySchemes:
    tossBearer:
      type: http
      scheme: bearer
      description: |
        두 종류 토큰(§8): ① `anon:{hash}` — getAnonymousKey() 익명 키(현재 기본, 사업자 등록 전) ② 토스 로그인 accessToken — 서버가 토스 login-me로 검증 후 userKey 식별. 어느 쪽이든 유저별 저장.
  responses:
    SettingNotFound:
      description: 통근 설정 미등록 — 온보딩으로 유도
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { code: SETTING_NOT_FOUND, message: 통근 설정이 없습니다 }
    InvalidRequest:
      description: 필드 검증 실패 — message에 상세
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { code: INVALID_REQUEST, message: 필드 검증 실패 }
    Unauthorized:
      description: 토스 토큰 없음/만료 — 클라이언트는 토스 재로그인
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { code: UNAUTHORIZED, message: 인증 실패 }
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: SETTING_NOT_FOUND }
        message: { type: string, example: 통근 설정이 없습니다 }
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, expiresInSeconds]
      properties:
        accessToken: { type: string, description: 토스 로그인 accessToken — 모든 요청의 Bearer (유효 1시간) }
        refreshToken: { type: string, description: 갱신 토큰 (유효 14일) — 클라이언트가 로컬 보관 }
        expiresInSeconds: { type: integer, format: int64, example: 3599 }
    StopType:
      type: string
      enum: [SEOUL_BUS, GYEONGGI_BUS, SUBWAY]
      description: SEOUL_BUS(stopId=arsId 5자리) / GYEONGGI_BUS(stopId=GBIS stationId) / SUBWAY(stopId=역명)
    DayOfWeek:
      type: string
      enum: [MON, TUE, WED, THU, FRI, SAT, SUN]
    CommuteRoute:
      type: object
      required: [id, label, enabled, setting]
      description: 통근 경로 (유저당 최대 5개) — FR-403·FR-405는 경로 단위로 적용
      properties:
        id: { type: string, description: 서버 발급 UUID (이관 행은 "migrated"), example: "550e8400-e29b-41d4-a716-446655440000" }
        label: { type: string, description: 1~16자, 유저 내 중복 불가, example: 출근 }
        enabled: { type: boolean, description: false = 이 경로 알림 일시 중지 }
        setting: { $ref: "#/components/schemas/CommuteSetting" }
    CommuteRouteRequest:
      type: object
      required: [label, setting]
      description: 생성·수정 요청 — id는 URL 경로로만 전달
      properties:
        label: { type: string, example: 퇴근 }
        enabled: { type: boolean, default: true }
        setting: { $ref: "#/components/schemas/CommuteSetting" }
    CommuteRoutesResponse:
      type: object
      required: [routes]
      properties:
        routes:
          type: array
          maxItems: 5
          items: { $ref: "#/components/schemas/CommuteRoute" }
    CommuteSetting:
      type: object
      required: [home, stops, walkMinutes, notificationMode, bufferMinutes, activeDays]
      properties:
        home:
          type: object
          description: 집 위치 위경도 (Granite getCurrentLocation 또는 지도 선택)
          required: [latitude, longitude]
          properties:
            latitude: { type: number, example: 37.5219 }
            longitude: { type: number, example: 126.9245 }
        stops:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/CommuteStop" }
        walkMinutes:
          type: integer
          description: 집→정류장 도보 시간(분). 유저 직접 입력 (FR-104)
          example: 8
        notificationMode:
          type: string
          enum: [FIXED, RECOMMENDED]
          description: ⓐ FIXED 정시 / ⓑ RECOMMENDED 추천 — 기본값 없음, 온보딩에서 선택
        fixedDepartureTime:
          type: string
          nullable: true
          description: '"HH:mm" — 모드 ⓐ일 때 필수, ⓑ면 null'
          example: "08:20"
        commuteWindow:
          type: object
          nullable: true
          description: 모드 ⓑ일 때 필수, ⓐ면 null
          properties:
            start: { type: string, example: "07:30" }
            end: { type: string, example: "09:00" }
        bufferMinutes:
          type: integer
          description: 여유 버퍼(분). 온보딩에서 유저가 반드시 선택 — 서버 기본값 강제 금지 (FR-302)
          example: 3
        activeDays:
          type: array
          items: { $ref: "#/components/schemas/DayOfWeek" }
          description: 알림 적용 요일 — 미포함 요일은 미발송 (FR-405)
    CommuteStop:
      type: object
      required: [type, stopId, displayName, routes]
      properties:
        type: { $ref: "#/components/schemas/StopType" }
        stopId: { type: string, example: "19284" }
        displayName: { type: string, example: 여의도환승센터 }
        routes:
          type: array
          minItems: 1
          items: { type: string }
          example: ["720", "261"]
    ArrivalsResponse:
      type: object
      required: [fetchedAt, realtimeAvailable, walkMinutes, arrivals]
      properties:
        fetchedAt:
          type: string
          format: date-time
          description: 서버가 공공 API에서 받아온 시각 — 화면에 "마지막 갱신 시각"으로 표시 (FR-501, NFR-07)
        realtimeAvailable:
          type: boolean
          description: false면 공공 API 장애/정보 없음 — "실시간 정보 없음" 명시 (NFR-03)
        walkMinutes: { type: integer, example: 8 }
        arrivals:
          type: array
          items: { $ref: "#/components/schemas/Arrival" }
    Arrival:
      type: object
      properties:
        stopDisplayName: { type: string, example: 여의도환승센터 }
        routeName: { type: string, example: "720" }
        direction:
          type: string
          nullable: true
          description: 방면 — 지하철 상행/하행 등, 버스는 null
          example: 상행
        secondsToArrival:
          type: integer
          nullable: true
          description: 도착까지 남은 초. null = 실시간 정보 없음 — 아는 척 금지
          example: 540
        remainingStops: { type: integer, nullable: true, example: 3 }
        isExpress:
          type: boolean
          nullable: true
          description: 지하철 급행 여부(버스는 null) — 급행/일반 구분 표시 (FR-203)
        boardable:
          type: boolean
          description: secondsToArrival ≥ 도보 시간이면 true (FR-303/304)
        status:
          type: string
          enum: [RELAXED, HURRY, MISSED]
          description: 여유/서두르세요/놓침→다음 차 (FR-502). 색상 외 수단으로도 구분 (NFR-06)
        rawMessage: { type: string, example: "9분후[3번째 전]" }
    BoardingFeedbackRequest:
      type: object
      required: [result, notifiedDate]
      properties:
        result:
          type: string
          enum: [BOARDED, MISSED]
          description: 탔어요 / 놓쳤어요
        notifiedDate:
          type: string
          format: date
          description: 해당 출근일
          example: "2026-08-28"
    StopSearchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            type: object
            required: [type, stopId, displayName, subtitle]
            properties:
              type: { $ref: "#/components/schemas/StopType" }
              stopId: { type: string, example: "19284" }
              displayName: { type: string, example: 여의도환승센터 }
              subtitle:
                type: string
                description: 동명 정류장 구분용 보조 정보 (정류소 번호·행정구 / 호선). 항상 존재 — 없으면 빈 문자열(null/누락 금지)
                example: 서울 · 19284
    StopRoutesResponse:
      type: object
      required: [routes]
      properties:
        routes:
          type: array
          items:
            type: object
            required: [name]
            properties:
              name: { type: string, example: "720" }
              isExpress: { type: boolean, nullable: true }